backup_before_cleanup/requirements.txt
# Core dependencies
opencv-python>=4.10.0
numpy>=1.26.4
matplotlib>=3.9.0
ezdxf>=1.3.0
shapely>=2.0.2

# Web interface
flask>=3.0.3
werkzeug>=3.0.0
streamlit>=1.34.0
pandas>=2.0.0

# Visualization and animation
plotly>=5.24.0

# CLI and utilities
rich>=13.7.1
pydantic>=2.7.0
click>=8.1.0

# Optional: For advanced image processing
pillow>=10.4.0
scikit-image>=0.24.0

# AI Integration
openai>=1.0.0
pyyaml>=6.0
requests>=2.25.0
python-dotenv>=1.0.1

# Evaluation and Testing
seaborn>=0.12.0

# Development and testing
pytest>=6.0.0
pytest-cov>=3.0.0

# Security and validation
python-magic>=0.4.27

# Authentication and Security
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0
flask-limiter>=3.5.0

# Database
SQLAlchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Async Processing
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0

# API and Web
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Nesting and Geometry
rectpack>=0.2.2
svgpathtools>=1.6.1

# PDF parsing & extraction
PyMuPDF>=1.24.8

# Monitoring and Observability
prometheus-client>=0.19.0
structlog>=23.2.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-jaeger-thrift>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-celery>=0.41b0
opentelemetry-instrumentation-sqlalchemy>=0.41b0
opentelemetry-instrumentation-redis>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
psutil>=5.9.0

# Testing and Quality
pytest-asyncio>=0.21.0
httpx>=0.25.0
faker>=20.0.0
